reecriture avec -Weverything part 1

+++ ft_itoa.c
@@ -6,7 +6,7 @@
 /*   By: amehmeto <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/30 18:32:27 by amehmeto          #+#    #+#             */
-/*   Updated: 2017/01/06 20:35:24 by amehmeto         ###   ########.fr       */
+/*   Updated: 2017/05/07 00:10:34 by amehmeto         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,12 +15,12 @@
 char	*ft_itoa(int n)
 {
 	long	m;
-	int		size;
+	long	size;
 	char	*rslt;
 
 	m = n;
 	size = ft_nbdigits(n) + (m < 0);
-	if (!(rslt = (char *)ft_strnew(size)))
+	if (!(rslt = (char *)ft_strnew((size_t)size)))
 		return (NULL);
 	if (m < 0)
 	{
