+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: amehmeto <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/23 01:44:24 by amehmeto          #+#    #+#             */
-/*   Updated: 2017/01/03 05:19:34 by amehmeto         ###   ########.fr       */
+/*   Updated: 2017/05/07 00:23:38 by amehmeto         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ int		ft_strncmp(const char *s1, const char *s2, size_t n)
 	while (s1[i] || s2[i])
 	{
 		if (s1[i] != s2[i] && i < n)
-			return (unsigned int)((unsigned char)s1[i] - (unsigned char)s2[i]);
+			return ((unsigned char)s1[i] - (unsigned char)s2[i]);
 		i++;
 	}
 	return (0);
