+++ ft_strsplit.c
@@ -6,7 +6,7 @@
 /*   By: amehmeto <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/30 15:59:10 by amehmeto          #+#    #+#             */
-/*   Updated: 2016/12/06 03:24:46 by amehmeto         ###   ########.fr       */
+/*   Updated: 2017/05/07 00:14:18 by amehmeto         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 char	**ft_strsplit(char const *s, char c)
 {
 	char			**rslt;
-	unsigned int	start;
+	size_t			start;
 	size_t			i;
 	size_t			j;
 
@@ -33,7 +33,7 @@ char	**ft_strsplit(char const *s, char c)
 			while (s[i] != c && s[i] != '\0')
 				i++;
 			if (s[start])
-				rslt[j++] = ft_strsub(s, start, i - start);
+				rslt[j++] = ft_strsub(s, (unsigned int)start, i - start);
 		}
 		rslt[j] = NULL;
 		return (rslt);
